# devintel-backend/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "DevIntel AI",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/repos": {
      "get": {
        "tags": [
          "Repositories"
        ],
        "summary": "List Repositories",
        "description": "List repositories (personal).",
        "operationId": "list_repositories_api_v1_repos_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Skip"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RepositoryListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Repositories"
        ],
        "summary": "Create Repository",
        "description": "Add a repository.",
        "operationId": "create_repository_api_v1_repos_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RepositoryCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RepositoryResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/repos/{repository_id}/pulls": {
      "get": {
        "tags": [
          "PR Review"
        ],
        "summary": "List Pull Requests",
        "description": "List pull requests for a repository from GitHub.",
        "operationId": "list_pull_requests_api_v1_repos__repository_id__pulls_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "repository_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Repository Id"
            }
          },
          {
            "name": "state",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "open",
              "title": "State"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "per_page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 30,
              "title": "Per Page"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PullRequestListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/pr-review": {
      "post": {
        "tags": [
          "PR Review"
        ],
        "summary": "Review Pull Request",
        "description": "Review a pull request using AI.",
        "operationId": "review_pull_request_api_v1_pr_review_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PRReviewRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PRReviewResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "IndexingStatus": {
        "type": "string",
        "enum": [
          "pending",
          "indexing",
          "complete",
          "failed"
        ],
        "title": "IndexingStatus"
      },
...
      "PRReviewRequest": {
        "properties": {
          "repository_id": {
            "type": "string",
            "format": "uuid",
            "title": "Repository Id",
            "description": "Repository ID"
          },
          "pull_request_diff": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pull Request Diff",
            "description": "PR diff content (optional if pr_number provided)"
          },
          "pr_number": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pr Number",
            "description": "GitHub PR number"
          },
          "pr_title": {
            "type": "string",
            "title": "Pr Title",
            "description": "PR title"
          },
          "pr_description": {
            "type": "string",
            "title": "Pr Description",
            "description": "PR description",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "repository_id",
          "pr_title"
        ],
        "title": "PRReviewRequest",
        "description": "PR review request schema."
      },
      "PRReviewResponse": {
        "properties": {
          "summary": {
            "type": "string",
            "title": "Summary",
            "description": "Overall assessment of the PR"
          },
          "potential_issues": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Potential Issues",
            "description": "Potential issues found"
          },
          "refactoring_suggestions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Refactoring Suggestions",
            "description": "Refactoring suggestions"
          },
          "security_warnings": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Security Warnings",
            "description": "Security concerns"
          },
          "performance_notes": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Performance Notes",
            "description": "Performance considerations"
          }
        },
        "type": "object",
        "required": [
          "summary"
        ],
        "title": "PRReviewResponse",
        "description": "PR review response schema."
      },
...
      "PullRequestListResponse": {
        "properties": {
          "pulls": {
            "items": {
              "$ref": "#/components/schemas/PullRequestResponse"
            },
            "type": "array",
            "title": "Pulls"
          },
          "repository_id": {
            "type": "string",
            "format": "uuid",
            "title": "Repository Id"
          }
        },
        "type": "object",
        "required": [
          "pulls",
          "repository_id"
        ],
        "title": "PullRequestListResponse",
        "description": "Pull request list response schema."
      },
      "PullRequestResponse": {
        "properties": {
          "number": {
            "type": "integer",
            "title": "Number"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "state": {
            "type": "string",
            "title": "State"
          },
          "author": {
            "type": "string",
            "title": "Author"
          },
          "author_avatar": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Author Avatar"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "title": "Updated At"
          },
          "additions": {
            "type": "integer",
            "title": "Additions"
          },
          "deletions": {
            "type": "integer",
            "title": "Deletions"
          },
          "url": {
            "type": "string",
            "title": "Url"
          }
        },
        "type": "object",
        "required": [
          "number",
          "title",
          "state",
          "author",
          "created_at",
          "updated_at",
          "additions",
          "deletions",
          "url"
        ],
        "title": "PullRequestResponse",
        "description": "Pull request response schema."
      },
...
      "RepositoryCreate": {
        "properties": {
          "repo_name": {
            "type": "string",
            "title": "Repo Name"
          },
          "full_name": {
            "type": "string",
            "title": "Full Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "url": {
            "type": "string",
            "title": "Url"
          },
          "stars": {
            "type": "integer",
            "title": "Stars",
            "default": 0
          },
          "language": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Language"
          },
          "default_branch": {
            "type": "string",
            "title": "Default Branch",
            "default": "main"
          }
        },
        "type": "object",
        "required": [
          "repo_name",
          "full_name",
          "url"
        ],
        "title": "RepositoryCreate"
      },
...
      "RepositoryListResponse": {
        "properties": {
          "repositories": {
            "items": {
              "$ref": "#/components/schemas/RepositoryResponse"
            },
            "type": "array",
            "title": "Repositories"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          }
        },
        "type": "object",
        "required": [
          "repositories",
          "total"
        ],
        "title": "RepositoryListResponse"
      },
      "RepositoryResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "user_id": {
            "type": "integer",
            "title": "User Id"
          },
          "github_repo_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Github Repo Id"
          },
          "full_name": {
            "type": "string",
            "title": "Full Name"
          },
          "repo_name": {
            "type": "string",
            "title": "Repo Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Url"
          },
          "stars": {
            "type": "integer",
            "title": "Stars",
            "default": 0
          },
          "language": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Language"
          },
          "default_branch": {
            "type": "string",
            "title": "Default Branch"
          },
          "is_private": {
            "type": "boolean",
            "title": "Is Private"
          },
          "indexing_status": {
            "$ref": "#/components/schemas/IndexingStatus"
          },
          "last_indexed_commit": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Indexed Commit"
          }
        },
        "type": "object",
        "required": [
          "id",
          "user_id",
          "full_name",
          "repo_name",
          "default_branch",
          "is_private",
          "indexing_status"
        ],
        "title": "RepositoryResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}